Refactor controller bootstrapping, centralise setup
