+++ controllers/about/index.js
@@ -1,22 +1,11 @@
 'use strict';
-const express = require('express');
-
-const routerSetup = require('../setup');
 const routeCommon = require('../common');
 const ebulletinRoute = require('./ebulletin');
 const seniorManagementRoute = require('./seniorManagement');
 const boardRoute = require('./boardRoute');
 const { shouldServe } = require('../../modules/pageLogic');
 
-const router = express.Router();
-
-module.exports = (pages, sectionPath, sectionId) => {
-    routerSetup({
-        router,
-        pages,
-        sectionId
-    });
-
+module.exports = (router, pages, sectionPath, sectionId) => {
     ebulletinRoute.init({
         router: router,
         routeConfig: pages.ebulletin
