+++ controllers/funding/index.js
@@ -1,8 +1,5 @@
 'use strict';
 
-const express = require('express');
-
-const routerSetup = require('../setup');
 const routeCommon = require('../common');
 
 const landingPageRoute = require('./funding');
@@ -10,15 +7,7 @@ const materialsRoute = require('./materials');
 const programmesRoute = require('./programmes');
 const tenKRoutes = require('./10k');
 
-const router = express.Router();
-
-module.exports = (pages, sectionPath, sectionId) => {
-    routerSetup({
-        router,
-        pages,
-        sectionId
-    });
-
+module.exports = (router, pages, sectionPath, sectionId) => {
     /**
      * Funding landing page
      */
