+++ controllers/research/index.js
@@ -1,18 +1,7 @@
 'use strict';
-const express = require('express');
-
-const routerSetup = require('../setup');
 const routeCommon = require('../common');
 
-const router = express.Router();
-
-module.exports = (pages, sectionPath, sectionId) => {
-    routerSetup({
-        router,
-        pages,
-        sectionId
-    });
-
+module.exports = (router, pages, sectionPath, sectionId) => {
     routeCommon.init({
         router,
         pages,
