+++ controllers/toplevel/index.js
@@ -1,15 +1,10 @@
 'use strict';
-
-const express = require('express');
 const config = require('config');
 const { sortBy } = require('lodash');
 const { body, validationResult } = require('express-validator/check');
 const moment = require('moment');
 const Raven = require('raven');
 
-const router = express.Router();
-
-const routerSetup = require('../setup');
 const routeCommon = require('../common');
 const surveyService = require('../../services/surveys');
 const contentApi = require('../../services/content-api');
@@ -24,13 +19,7 @@ const searchRoute = require('./search');
 const feedbackRoute = require('./feedback');
 const legacyPages = require('./legacyPages');
 
-module.exports = (pages, sectionPath, sectionId) => {
-    routerSetup({
-        router,
-        pages,
-        sectionId
-    });
-
+module.exports = (router, pages, sectionPath, sectionId) => {
     /**
      * Robots / Sitemap
      */
